NP-45338 Move candidate actions to own component
